Remove the bit of VC 64bit compilation warning
git-svn-id: https://svn.java.net/svn/j3d-core~svn/trunk@715 ba19aa83-45c5-6ac9-afd3-db810772062c

+++ src/native/ogl/Attributes.c
@@ -485,7 +485,7 @@ void JNICALL Java_javax_media_j3d_NativePipeline_setLightEnables(
 #endif
 
     for (i=0; i<nlights; i++) {
-	if (enable_mask & (1L<<i)) {
+      if (enable_mask & ((long)(1<<i))) {
 	    glEnable(GL_LIGHT0 + i);
 	}
 	else {
